requirement.txt
streamlit
requests
beautifulsoup4
tqdm
langchain-groq
googlesearch-python
langchain_community
python-dotenv
